Refactored delegateConsumerRecord() (which was an absolute behemoth) from LeaderFollowerStoreIngestionTask / ActiveActiveStoreIngestionTask into StorePartitionDataReceiver. 🧹
